requirements.txt
scipy==0.18.0
dask==0.11.0
decorator==4.0.10
networkx==1.11
numpy==1.11.1
oauthlib==2.0.0
Pillow==3.3.1
requests==2.11.1
requests-oauthlib==0.6.2
scikit-image==0.12.3
six==1.10.0
toolz==0.8.0
twython==3.4.0
